Feature/payment service 31

+++ Pendo.Database/Tables/payment/UserBalance.sql
@@ -7,8 +7,8 @@ Description: Creates User's Balance table.
 CREATE TABLE [payment].[UserBalance]
 (
   [UserId] UNIQUEIDENTIFIER NOT NULL PRIMARY KEY,
-  [Pending] FLOAT NOT NULL DEFAULT 0,
-  [NonPending] FLOAT NOT NULL DEFAULT 0,
+  [Pending] DECIMAL(18,8) NOT NULL DEFAULT 0,
+  [NonPending] DECIMAL(18,8) NOT NULL DEFAULT 0,
   [UpdatedDate] DATETIME2 NOT NULL DEFAULT GETUTCDATE(),
 
   CONSTRAINT FK_User_UserId FOREIGN KEY ([UserId])
